Fix highlighting for $# in the shell lexer.
If '$' is directly followed by '#' the variable ends right there. This is especially important at the end of a string, as the closing quote would be lexed as part of the variable instead, but was wrong in every other case as well.

Fixed by simply removing the '#?' and letting the following '.' eat it.